UI/v2 new UI (#19)
* chore: Refactor layout components for improved code organization and consistency

* chore: Refactor layout components for improved code organization and consistency

* feat: Add .env to .gitignore and implement sunset and sunrise data fetching in YearlyProgressWebApp

* Added Sunset and Sunrise

* Fix: Moved Sunrise and Sunset widget to client side

* Temp fix

* idk

* Moved to asking user for location permission

* Fix

* chore: Update YearlyProgressWidget transition duration for smoother animation

* Changed the api to sunrise

* Removed the use of OpenWeatherMap API Key

* Fix date issue

* feat: Update YearlyProgressSunsetSunriseWidget to use separate Daylight and Night widgets

The YearlyProgressSunsetSunriseWidget component has been updated to improve the user experience by separating the functionality into two separate widgets: YearlyProgressDaylightWidget and YearlyProgressNightWidget. This change allows for more flexibility and customization in displaying the sunrise and sunset times.

This commit also includes the removal of the OpenWeatherMap API key and the use of the sunrise-sunset.io API for fetching the sunset and sunrise times. Additionally, the date issue has been fixed, and the transition duration of the YearlyProgressWidget has been updated for smoother animation.

Fixes #123

* Fix: Wrong sunset and sunrise calculation for night time

* feat: Separate night time functionality into YearlyProgressNightWidget

* Separate night time functionality into YearlyProgressNightWidget

* feat: Update YearlyProgressWebApp to include v0.04 and add location permission request

The YearlyProgressWebApp component has been updated to include version v0.04 and a location permission request for the Daylight and Night Widget. This change improves the user experience by displaying the correct version number and requesting permission to access the user's location for accurate sunrise and sunset times.

Fixes #123

* feat: Update layout.tsx and eye-care.json, and remove unused route.ts

This commit updates the layout.tsx file by removing the import of the Footer component, improving the page load performance by reducing unnecessary dependencies.

In the eye-care.json file, the "about" description for the Eye Care app has been updated to include a period at the end of the sentence, improving the clarity of the text.

The route.ts file has been deleted as it is no longer needed, removing unnecessary code from the project.

* feat: Improve IP retrieval in route.ts

Refactor the GET function in route.ts to improve IP retrieval. Instead of relying solely on the request.ip property, the code now checks for the 'x-real-ip' header and falls back to the 'x-forwarded-for' header if necessary. This change ensures more accurate IP detection.

Fixes #456

* Added ip based laction as well

* feat: Improve IP retrieval in YearlyProgressDaylightWidget

* feat: Improve IP retrieval in YearlyProgressDaylightWidget

* feat: Improve IP retrieval in YearlyProgressDaylightWidget

* feat: Improve IP retrieval in YearlyProgressDaylightWidget

* Update eye-care.json with long description and add location permission message

* feat: Update page links for Yearly Progress and Eye Care

Update the page links in the Home component to point to the specific sections on the Yearly Progress and Eye Care pages. This change improves the user experience by directing users to the relevant information directly.

* Testing New About Page UI

* Update image file extensions to webp for better performance

* fix: Update member cards UI in About Us page