refactor: Update layout and eye-care.json, improve IP retrieval, and remove unused code
This commit refactors the layout.tsx file to improve page load performance by removing the import of the Footer component, reducing unnecessary dependencies.

In the eye-care.json file, the "about" description for the Eye Care app has been updated to include a period at the end of the sentence, improving the clarity of the text.

The route.ts file has been deleted as it is no longer needed, removing unnecessary code from the project.

Additionally, the IP retrieval has been improved in the route.ts and YearlyProgressDaylightWidget files to ensure more accurate detection.

Fixes #456